docker token nginx

+++ src/admin/rest/api.js
@@ -1,10 +1,13 @@
+const { exception } = require('console');
 const util = require('util');
 const exec = util.promisify(require('child_process').exec);
+const { makeid } = require('../../utils')
 
 const NAAS_TOKEN = process.env.NAAS_AUTH_TONE || 'DONT_USE_THIS_TOKEN'
 
 class Api {
-  constructor (rest) {
+  constructor (rest, tokens) {
+    this.tokens = tokens
     this.rest = rest
 
     this.NAAS_TCP_LOWERBOUND = parseInt(process.env.NAAS_TCP_LOWERBOUND || 9000)
@@ -12,6 +15,7 @@ class Api {
     this.NAAS_UDP_LOWERBOUND = parseInt(process.env.NAAS_UDP_LOWERBOUND || 59000)
     this.NAAS_UDP_UPPERBOUND = parseInt(process.env.NAAS_UDP_UPPERBOUND || 65000)
     this.NAAS_IMAGE = process.env.NAAS_IMAGE || 'm1k1o/neko:latest'
+    this.NAAS_NETWORK = process.env.NAAS_NETWORK || 'naas'
 
     this.rest.get('/api/test', this.mep(this.test))
     this.rest.get('/api/rooms', this.mep(this.listDockerContainers))
@@ -24,9 +28,10 @@ class Api {
   checkToken (req) {
     const auth = req.headers.authorization || ""
     const token = auth.split('Bearer ')[1]
-    if (token !== NAAS_TOKEN) {
+    if (!this.tokens[token]) {
       throw new Error( `Not authorized, expected  ${NAAS_TOKEN} got ${token} from ${auth}`)
     }
+    return { $id: token, ...this.tokens[token] }
   }
 
   mep(ep) {
@@ -34,17 +39,18 @@ class Api {
     const oThis = this
     return (async function (req, res) {
       try {
-      oThis.checkToken(req)
+      const $token = oThis.checkToken(req)
       res.ok(await ep({
           ...req.params,
           ...req.body,
           $req: req,
           $res: res,
-          $ctx: this
+          $ctx: this,
+          $token
         }))
       } catch (ex) {
         console.error('Error processing request', req.url, req.body, req.headers, ex)
-        res.internalServerError(ex)
+        res.internalServerError({ error: "" + ex })
       }
     }).bind(this)
   }
@@ -54,7 +60,7 @@ class Api {
   }
 
   async getMappedPorts () {
-    const { stdout, stderr } = await exec('docker ps --format "{{.Ports}}"')
+    const { stdout, stderr } = await exec(`docker ps --format "{{.Ports}}"`)
     /** stdout
      * 0.0.0.0:59101-59200->59101-59200/udp, 0.0.0.0:9081->8080/tcp
      * 0.0.0.0:59000-59100->59000-59100/udp, 0.0.0.0:9080->8080/tcp
@@ -117,16 +123,54 @@ class Api {
     throw new Error('No empty port found')
   }
 
-  async listDockerContainers () {
-    const { stdout, stderr } = await exec('docker ps --format "{{.Names}}"')
+  async listDockerContainers ({ $token }) {
+    const cmd = `docker ps -a --filter "label=naas.token=${$token.$id}" --format "{{.Names}}"`
+    console.log(cmd)
+    const { stdout, stderr } = await exec(cmd)
     if (stderr) {
       throw new Error(stderr)
     }
     return stdout.split('\n')
-      .filter(n => n.startsWith('neko_'))
+            .filter(l => l.length !== 0)
   }
 
-  async createNeko({ maxMembers, image }) {
+  async newContainerName ({ namePattern, $token }) {
+    namePattern = namePattern || 'xxx-xxx-xxxx'
+    const names = await this.listDockerContainers({ $token })
+    let name = makeid(namePattern)
+    let attemps = 5
+    while(attemps && names.indexOf(name) !== -1) {
+      name = makeid(namePattern)
+      attemps--
+    }
+    if (attemps === 0) {
+      throw new Error('No valid container name found')
+    }
+    return name
+  }
+
+  async canCreateNeko({ $token, maxMembers }) {
+    const current = await this.listDockerContainers({ $token })
+    if (current.length >= $token.concurrent) {
+      throw new Error('Token reached max number of running rooms')
+    }
+    if (maxMembers > $token.maxMembers) {
+      throw new Error('maxMembers error')
+    }
+    return true
+  }
+
+  async createNeko({
+    name,
+    namePattern,
+    maxMembers,
+    image,
+    password,
+    adminPassword,
+    $token
+  }) {
+    maxMembers = maxMembers || 2
+    await this.canCreateNeko({ $token, maxMembers })
     const port = await this.getFreePort({
       protocol: 'tcp',
       lowerBound: this.NAAS_TCP_LOWERBOUND,
@@ -136,11 +180,13 @@ class Api {
     const udpPort = await this.getFreePort({
       protocol: 'udp',
       lowerBound: this.NAAS_UDP_LOWERBOUND,
-      count: maxMembers || 10,
+      count: maxMembers,
       upperBound: this.NAAS_UDP_UPPERBOUND
     })
     const epr = `${udpPort}-${udpPort + maxMembers}`
-    const name = `neko_${port}`
+    if (!name) {
+      name = await this.newContainerName({ namePattern, $token })
+    }
     image = image || this.NAAS_IMAGE
 
     const tpl = `docker run -d \
@@ -148,21 +194,36 @@ class Api {
       -p ${port}:8080 \
       -p ${epr}:${epr}/udp \
       -e NEKO_EPR=${epr} \
+      -e NEKO_PASSWORD_ADMIN=${adminPassword || makeid('xxx-xxx-xxx')} \
+      -e NEKO_PASSWORD=${password || makeid('xxx-xxx-xxx')} \
+      --network ${this.NAAS_NETWORK} \
+      --label naas.token=${$token.$id} \
       --name ${name} \
       ${image}`
-    const { stdout, stderr } = await exec(tpl)
-    return {
-      port,
-      epr,
-      name,
-      image,
-      tpl,
-      stdout,
-      stderr
+
+    try {
+      const { stdout, stderr } = await exec(tpl)
+      return {
+        port,
+        epr,
+        name,
+        image,
+        adminPassword,
+        password,
+        tpl,
+        stdout
+      }
+    } catch (error) {
+      await this.deleteContainer({ name })
+      return { error }
     }
   }
 
-  async deleteContainer({ name }) {
+  async deleteContainer({ name, $token }) {
+    const current = await this.listDockerContainers({ $token })
+    if (current.indexOf(name) === -1) {
+      throw new Error('Container not found')
+    }
     const { stdout, stderr } = await exec(`docker rm -f ${name}`)
     return { status: stderr ? stderr : stdout.trim() === name }
   }
